Working on buttons and enemy unit
Player unit cannot move through enemy units, when moving, the highlight does not show on the tile an enemy is standing on. When attacking, the tile the enemy is standing on is included in the highlight. Added an attack damage stat to units. After pressing the attack button, the attack range is highlighted, and you can select an enemy within that range. After selecting, it prints how much damage the unit dealt. (will probably be a random number generator)