Progress on attacking, added item types, rearranged functions
Organized some things differently. Instead of the damage being written out in the update function of Unit2D, I made a separate function called DamageThis, and I tried to make the Unit2D script more generalized, so we can use it for the enemy units too. I have rearranged Tilemap2D so that the Start and Update functions are at the bottom of the script, so they are easier to find. I also tried to group the functions in some sort of order of what they do. I haven't done much with the item button yet.